Module pmfp_collections::archive_stack

High-gauge Stack

a stack implemented with both persistent and mutable components

insert in low-const O(1) - rare reallocations

copy in high-const O(1) - short array copy then pointer

The archive stack is implemented as a persistent stack of vectors. Use archive() to save the current vector as the first entry in the persistent linked list.

Structs

AStack
Archive Stack

Iter