Struct pmfp_collections::archive_stack::AStack

pub struct AStack<E: Clone, M: Clone> { /* fields omitted */ }

Archive Stack

Operations on a mutable vector, with a persistent stack of vectors available. pop() will open archives, but pushing to the archive requires the archive() method.

Parametric over elements, E, and metadata M, that will be included with archived vectors. Metadata is currently only available if the archive is accessed with next_archive()

Methods

impl<E: Clone, M: Clone> AStack<E, M>

fn new() -> Self

new AStack with a new vector as current stack

fn with_capacity(capacity: usize) -> Self

new AStack with a new pre-allocated vector as current stack

fn is_empty(&self) -> bool

whether or not the AStack has any data, including archived data

fn len(&self) -> usize

the total item count of the AStack, including archived data

fn active_len(&self) -> usize

the item count of the "fast" mutable vector outside of the archive

fn push(&mut self, elm: E)

push a element to the "fast" vector outside of the archive

fn pop(&mut self) -> Option<E>

remove and return the element at the top of the stack, even if it is within the archive.

archive metadata is lost when opening an archive in this way

fn next_archive(&mut self) -> Option<(Vec<E>, Option<M>)>

Exposes the last data archived, returning the prior active vector and the metadata stored with the archive

fn active_data(&self) -> &Vec<E>

peeks at the entire active vector

fn peek(&self) -> Option<&E>

peeks at the element at the top of the stack, even if it is within the archive.

peeking into the archive does not open it, so it will retain any associated metadata

fn archive(&mut self, new_meta: M) -> bool

push the entire active vector into the archive, along with associated metadata return false if the active vector was empty. In this case, no archive will happen and the metadata will be unused

fn archive_with_capacity(&mut self, new_meta: M, capacity: usize) -> bool

push the entire active vector into the archive, providing a capacity for the new active vector