Struct pmfp_collections::inc_gauged_raz::RazTree

pub struct RazTree<E: 'static + Debug + Clone + Eq + Hash> { /* fields omitted */ }

Tree form of a RAZ

used between refocusing, and for running global algorithms

Methods

impl<E: Debug + Clone + Eq + Hash + 'static> RazTree<E>

fn len(&self) -> usize

the number if items in the sequence

fn fold_up<I, R, B>(self, init: Rc<I>, bin: Rc<B>) -> Option<R> where R: 'static + Eq + Clone + Hash + Debug,
I: 'static + Fn(&E) -> R,
B: 'static + Fn(R, R) -> R

Runs an binary function over the sequence data

This is calculated from data in leaves of a tree structure, so the operation must be associative. Returns None if there are no elements.

fn fold_lr<A, B>(self, init: A, bin: Rc<B>) -> A where A: 'static + Eq + Clone + Hash + Debug, B: 'static + Fn(A, &E) -> A

left-to-right memoized fold

fn map<R, F>(self, f: Rc<F>) -> RazTree<R> where R: 'static + Eq + Clone + Hash + Debug, F: 'static + Fn(&E) -> R

returns a new tree with data mapped from the old tree

fn focus(self, pos: usize) -> Option<Raz<E>>

focus on a location in the sequence to begin editing.

0 is before the first element. This will return None if pos is larger than the number of elements.