Function pmfp_collections::level_tree::good_levels

pub fn good_levels<E>(tree: &Tree<E>) -> bool

Use good_levels to verify level consistancy when debugging

This is an O(n) operation, so it shouldn't be used in release mode

use pmfp_collections::level_tree::{self,Tree};

let tree = Tree::new(4,(),None,Tree::new(1,(),None,None)).unwrap();
debug_assert!(level_tree::good_levels(&tree),"this section of code has a problem");

checks that the levels of the tree follow the convention of non-increasing to the left branch and decreasing to the right branch

also prints the levels of the failing trees and branchs