Struct pmfp_collections::level_tree::Tree

pub struct Tree<E> { /* fields omitted */ }

A persistent tree with stable, internally defined structure

Methods

impl<E> Tree<E>

fn new(level: u32,
data: E,
l_branch: Option<Tree<E>>,
r_branch: Option<Tree<E>>)
-> Option<Tree<E>>

build a new tree from components, return None if levels are inconsistent

fn level(&self) -> u32

peek at the level of the root of this tree

fn l_tree(&self) -> Option<Tree<E>>

obtain the left subtree if it exists

fn r_tree(&self) -> Option<Tree<E>>

obtain the right subtree if it exists

fn peek(&self) -> &E

peek at the data contained at the top node of the tree

this functionality is also available through Deref

fn fold_up<R, F>(&self, node_calc: &mut F) -> R where F: FnMut(Option<R>, &E, Option<R>) -> R