Module pmfp_collections::split_btree_cursor

Structs

Cursor
tree cursor, centered on a node of the underlying persistent tree

Tree
A persistant tree, for use with a cursor. This tree is as general as possible to be of arbitrary use.

Enums

Force
cursor movement qualifier

T
deconstruction pattern

Traits

DeriveTreeUpdate
marker that allows a default implementation of TreeUpdate if the data is also Clone

TreeUpdate
Used for updating data when the tree is mutated

Functions

gen_level
generate a random level appropriate for a balanced binary tree

Type Definitions

Level
A tree with the cursor here needs "levels", which are non-decreasing from the root of the tree to the leaves.