Struct pmfp_collections::split_btree_cursor::Cursor

pub struct Cursor<E: TreeUpdate> { /* fields omitted */ }

tree cursor, centered on a node of the underlying persistent tree

A cursor allows exploration of the underlying tree by following links to branches or up towards root. This structure is optimised for splitting and combining trees in a cannonical way based on the levels of the nodes; that is, trees with the same levels will have the same structure, regaurdless of order of operations.

Many operations allow structural mutation of the underlying tree.

Methods

impl<E: TreeUpdate> Cursor<E>

fn new() -> Self

creates a new cursor, to an empty underlying tree

fn with_depth(capacity: usize) -> Self

creates a new cursor, with expected depth of underlying tree

fn split(self) -> (Cursor<E>, Tree<E>, Cursor<E>)

Returns the node the cursor is focused on as a tree, plus two cursors containing every node to the left and right, focused on at the two branches of the returned tree

fn join(l_cursor: Self, level: Level, data: E, r_cursor: Self) -> Self

makes a new cursor at the given data, between the trees of the other cursors

The update() method of the data type will be called, with the data parameter passed here as the old_data to that method (along with joined branches).

fn at_tree(&self) -> Tree<E>

copies the focused node as a tree

This is a persistent tree, so copies are Rc clones

fn left_tree(&self) -> Option<Tree<E>>

copies the left branch of the focused node

fn right_tree(&self) -> Option<Tree<E>>

copies the right branch of the focused node

fn peek(&self) -> Option<&E>

peek at the data of the focused tree node

fn peek_level(&self) -> Option<Level>

peek at the level of the focused tree node

fn down_left_force(&mut self, force: Force) -> bool

move the cursor into the left branch, returning true if successful use the Force enum to determine the type of movement

fn down_left(&mut self) -> bool

move the cursor to the left branch, without entering an empty branch

fn down_right_force(&mut self, force: Force) -> bool

move the cursor into the right branch, returning true if successful use the Force enum to determine the type of movement

fn down_right(&mut self) -> bool

move the cursor to the right branch, without entering an empty branch

fn up(&mut self) -> bool

move the cursor up towards the root of the underlying persistent tree

If the tree has been changed, the update() method of the tree's data type will be called as a new persistent node is created.