Struct pmfp_collections::split_btree_cursor::Tree

pub struct Tree<E>(_);

A persistant tree, for use with a cursor. This tree is as general as possible to be of arbitrary use.

Methods

impl<E> Tree<E>

Generic tree for use with the tree cursor

fn empty() -> Self

create an empty branch

fn new(level: Level,
element: E,
left_branch: Tree<E>,
right_branch: Tree<E>)
-> Tree<E>

build a new tree from components

fn is_empty(&self) -> bool

whether the tree contains any structure

fn level(&self) -> Level

peek at the level of the root of this tree, an empty tree has level 0

fn peek(&self) -> Option<&E>

peek at the data contained at the top node of the tree

fn good_levels(&self) -> bool

for debugging, this is an O(n) operation debug_assert!(tree.good_levels(),"this section of code has a problem");

checks that the levels of the tree follow the convention of non-increasing to the left branch and decreasing to the right branch

also prints the levels of the first failing tree and branches