Trait pmfp_collections::trees::BinTree

pub trait BinTree<T> where Self: Sized + Clone {
    fn l_tree(&self) -> Option<Self>;
    fn r_tree(&self) -> Option<Self>;
    fn peek(&self) -> &T;
    fn fold_up<R, F>(&self, node_calc: &mut F) -> R where F: FnMut(&T, Option<R>, Option<R>) -> R;

    fn bin_make(data: T, l_branch: Option<Self>, r_branch: Option<Self>) -> Self { ... }
}

General trait for a persistent binary tree

All nodes are trees themselves, and all trees are immutable and shared. An owned tree is really a link to a shared node.

Required Methods

fn l_tree(&self) -> Option<Self>

get the left branch

fn r_tree(&self) -> Option<Self>

get the right branch

fn peek(&self) -> &T

get a reference to the data in this branch

fn fold_up<R, F>(&self, node_calc: &mut F) -> R where F: FnMut(&T, Option<R>, Option<R>) -> R

Perform a calculation recursively over all branches.

The function node_calc takes the data from the current branch and the result of a recursive call to the lower branches, if they exist.

Provided Methods

fn bin_make(data: T, l_branch: Option<Self>, r_branch: Option<Self>) -> Self

construct a new tree with existing nodes. The default implementation is a panic!() requesting use of a type constructor.